+++ src/Fable.Transforms/Python/Fable2Python.Transforms.fs
@@ -653,6 +653,7 @@ let transformObjectExpr
                             IsThisArgument = false
                             IsCompilerGenerated = true
                             Range = r
+                            IsInlineIfLambda = false
                         }
                 | e -> e
             )